backend/requirements.txt
fastapi==0.115.14
SQLAlchemy==2.0.41
passlib==1.7.4
python-jose==3.5.0
